move useRefUndoState to hooks directory
